Added tutor schedule and integrated TutorAddSubject with new navbar
